--- src/components/universalComponents/RecipesList/RecipesList.styled.js
@@ -1,64 +1,0 @@
-import styled from 'styled-components';
-
-export const List = styled.ul`
-  list-style: none;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 28px;
-
-  @media screen and (min-width: 768px) {
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: flex-start;
-    gap: 32px;
-  }
-
-  @media screen and (min-width: 1440px) {
-    gap: 14px;
-    row-gap: 100px;
-  }
-`;
-
-export const Recipe = styled.li`
-  overflow: hidden;
-  border-radius: 8px;
-  padding: 0;
-`;
-
-export const RecipeContainer = styled.div`
-  position: relative;
-`;
-
-export const Description = styled.div`
-  position: absolute;
-  left: 18px;
-  bottom: 24px;
-  padding: 16px;
-  width: 307px;
-  background: #ffffff;
-  border-radius: 8px;
-
-  @media screen and (min-width: 768px) {
-    left: 16px;
-    width: 300px;
-  }
-  @media screen and (min-width: 1440px) {
-    width: 268px;
-  }
-`;
-
-export const RecipeImg = styled.img`
-  width: 343px;
-  height: 323px;
-  object-fit: cover;
-  transition: ${props => props.theme.effects.mainEffect};
-
-  @media screen and (min-width: 768px) {
-    width: 336px;
-  }
-  @media screen and (min-width: 1440px) {
-    width: 299px;
-    gap: 14px;
-  }
-`;
